+++ include/SumCalculator.h
@@ -37,7 +37,7 @@ namespace WeatherAnalysis
 
 	boost::shared_ptr<Acceptor> itsAcceptor;
 	long itsCounter;
-	float itsSum;
+	double itsSum;
 
   }; // class SumCalculator
